Updated createpoll.js, polldata.js, closePollButton.js, voteModalPopupButton.js, modalPollSubmit.js, interactionCreate.js, ready.js, to allow for multiple concurrent polls
Created a save and load functionallity so that polls are stored on disk TODO Make limits on this to a certain number of concurrent polls so as to not have to big of storage file on disk.
Polls will be stored to disk on any updates. On bot start, polls will be loaded. This allows for poll data to not be lost if the bot is shutdown while a poll(s) is active